fix(ci,#8884): non-render detector gates errexit-safe + generalized ratchet
The four remaining detector gates carried the #8884 bash -e defect
(`out=$(cmd)` + `rc=$?`): under `bash -e` a non-zero detector aborts the
whole step before `rc=$?` is read, making the rc=1/rc=2 branches dead code
and the failure anonymous (the gate can never NAME the notebook it caught):

  - bare-cross-dir-load-gate.yml    (detect_bare_cross_dir_load, exit 0/1/2)
  - manifest-description-visuelle-gate.yml (detect_manifest_field, 0/1/2)
  - md-content-loss-gate.yml        (detect_md_content_loss, 0/1/2)
  - cell-order-gate.yml             (cell_order_ci, 0/1)

NOTE: cell-order-gate is a FOURTH non-render gate -- #8890's body cited 3
(md/bare/manifest). cell_order_ci lacks the `detect_` prefix the render-suite
ratchet matched, so it fell outside the earlier discovery too.

Fix: the `&& rc=0 || rc=$?` tail (membership in a &&/|| list is errexit-
exempt; `$?` reads the real detector exit code). NOT `|| true` (that makes
`$?` read 0 from `true` and permanently disarms the gate).

Generalized ratchet: replaces test_render_suite_gates.py (render-suite-only,
#8890) with test_detector_gate_errexit_safety.py, which auto-discovers EVERY
`*-gate.yml` that captures a detector exit into `out=$(python
scripts/notebook_tools/...)` -- so a gate added later with the bug is caught
automatically, render-suite or not. Covers all 9 known detector gates.

Proof: RED->GREEN verified -- reverting cell-order-gate.yml to the bare
`out=$(cmd)` form fails the ratchet with a named, actionable assertion;
restoring the fix passes (10/10). Bounded detector suite 268 passed.

See #8884 (root motif), #8890 (render-suite predecessor that anticipated
this follow-up), #8886 (degenerate-figure-gate original fix).
